e2e: rte: check the binary path only in local test
Only a well known subset of tests (RTE local) needs
the local exporter binary, so let's look for it and consume
it only for that subset instead for the whole suite.
