+++ test/e2e/rte/rte_e2e_suite_test.go
@@ -21,7 +21,6 @@ import (
 	"fmt"
 	"math/rand"
 	"os"
-	"os/exec"
 	"testing"
 	"time"
 
@@ -33,15 +32,12 @@ import (
 
 	e2etestns "github.com/k8stopologyawareschedwg/resource-topology-exporter/test/e2e/utils/namespace"
 	"github.com/openshift-kni/numaresources-operator/internal/objects"
-	"github.com/openshift-kni/numaresources-operator/test/utils/runtime"
 
 	_ "github.com/k8stopologyawareschedwg/resource-topology-exporter/test/e2e/rte"
 	_ "github.com/k8stopologyawareschedwg/resource-topology-exporter/test/e2e/topology_updater"
 )
 
 var (
-	BinaryPath string
-
 	randomSeed int64
 )
 
@@ -68,24 +64,4 @@ func TestRTE(t *testing.T) {
 
 var _ = BeforeSuite(func() {
 	By(fmt.Sprintf("Using random seed %v", randomSeed))
-
-	By("Finding the binaries path")
-
-	binPath, err := runtime.FindBinaryPath("exporter")
-	Expect(err).ToNot(HaveOccurred())
-	BinaryPath = binPath
-
-	By(fmt.Sprintf("Using the binary at %q", BinaryPath))
 })
-
-func expectExecutableExists(path string) {
-	cmdline := []string{
-		path,
-		"-h",
-	}
-
-	cmd := exec.Command(cmdline[0], cmdline[1:]...)
-	out, err := cmd.CombinedOutput()
-	ExpectWithOffset(1, err).ToNot(HaveOccurred())
-	ExpectWithOffset(1, out).ToNot(BeEmpty())
-}
